added changes on Qoute Generator and connect to firestore database
