+++ assets/json/rizzCard.json
@@ -1,22 +1,18 @@
 [
     {
       "id": 1,
-      "message": "Ikaw ang akong favorite subject, bisan unsa pa ang topic."
-    },
-    {
-      "id": 2,
       "message": "If you were a command, you'd be 'git commit -m “My perfect match”.'"
     },
     {
-      "id": 3,
+      "id": 2,
       "message": "You light up my world like syntax highlighting brightens my code."
     },
     {
-      "id": 4,
+      "id": 3,
       "message": "I may be a programmer, but you’re the one who’s coding my heart."
     },
     {
-      "id": 5,
+      "id": 4,
       "message": "You're the semicolon to my line of code—essential and completing my thoughts."
     }
   ]
